refactor: separate logic to different file for consistency
